+++ src/imgaug/cl/kernel/flip.cl
@@ -7,14 +7,14 @@ __kernel void flip_horizontal_planar(
 )
 {
 
-    unsigned short id_x = get_global_id(0);
-    unsigned short id_y = get_global_id(1);
-    unsigned short id_z = get_global_id(2);
+    int id_x = get_global_id(0);
+    int id_y = get_global_id(1);
+    int id_z = get_global_id(2);
     if (id_x >= width || id_y >= height || id_z >= channel) return;
 
-    unsigned short oPixIdx =   id_x + id_y * width + id_z * width * height;
+    int oPixIdx =   id_x + id_y * width + id_z * width * height;
 
-    unsigned short nPixIdx =   id_x + (height-1 - id_y) * width + id_z * width * height;
+    int nPixIdx =   id_x + (height-1 - id_y) * width + id_z * width * height;
 
 	output[nPixIdx] = input[oPixIdx];
 
@@ -28,14 +28,14 @@ __kernel void flip_vertical_planar(
     const unsigned short channel
 )
 {
-    unsigned short id_x = get_global_id(0);
-    unsigned short id_y = get_global_id(1);
-    unsigned short id_z = get_global_id(2);
+    int id_x = get_global_id(0);
+    int id_y = get_global_id(1);
+    int id_z = get_global_id(2);
     if (id_x >= width || id_y >= height || id_z >= channel) return;
 
-    unsigned short oPixIdx =   id_x + id_y * width + id_z * width * height;
+    int oPixIdx =   id_x + id_y * width + id_z * width * height;
 
-    unsigned short nPixIdx =   (width-1 - id_x) + id_y * width + id_z * width * height;
+    int nPixIdx =   (width-1 - id_x) + id_y * width + id_z * width * height;
 
 	output[nPixIdx] = input[oPixIdx];
 
@@ -50,14 +50,14 @@ __kernel void flip_bothaxis_planar(
 )
 {
 
-    unsigned short id_x = get_global_id(0);
-    unsigned short id_y = get_global_id(1);
-    unsigned short id_z = get_global_id(2);
+    int id_x = get_global_id(0);
+    int id_y = get_global_id(1);
+    int id_z = get_global_id(2);
     if (id_x >= width || id_y >= height || id_z >= channel) return;
 
-    unsigned short oPixIdx =   id_x + id_y * width + id_z * width * height;
+    int oPixIdx =   id_x + id_y * width + id_z * width * height;
 
-    unsigned short nPixIdx =   (width-1 - id_x) + (height-1 - id_y) * width + id_z * width * height;
+    int nPixIdx =   (width-1 - id_x) + (height-1 - id_y) * width + id_z * width * height;
 
 	output[nPixIdx] = input[oPixIdx];
